check for broker IP before connecting, add restart config option
The initMQTT will now check for a broker IP or host before attempting to make a connection.

A new config option is added to let the user have the module auto restart when an error occurs, or not. (Previously it would auto restart after 1 second, and this could make it difficult to troubleshoot or see what is going on.)